add TocItem.viewOnRelativePath, move MarkupParsingConfiguration to core
